pyproject.toml
[tool.poetry]
name = "olympus"
version = "0.1.0"
description = ""
authors = ["Christian Haack <christian.haack@icecube.wisc.edu>"]
readme = "README.md"
packages = [{ include = "olympus" }]

[tool.poetry.dependencies]
python = ">=3.10, <3.12"
numpy = ">=1.21.0,<2.0.0"
pandas = ">=1.4.0,<2.0.0"
matplotlib = ">3.4"
jax = ">0.4.0"
plotly = ">5.0.0"
awkward = ">=1.8"
pydantic = ">1.8.0"
pyarrow = ">9.0.0"
tqdm = ">4.64"
fennel-seed = ">1.3.0"
# TODO: Set to fixed version once conan~=1.33 is fixed
proposal = { git = "git@github.com:tudo-astroparticlephysics/PROPOSAL.git", "branch" = "master" }
hyperion = {git = "https://github.com/pone-software/hyperion.git"}
pone-ananke = {git = "https://github.com/pone-software/ananke.git", rev = "v0.2.2"}


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
